Remove testing and proxy helpers
Ref: https://github.com/beda-software/fhir-client-testing-tool/issues/7
