+++ packages/app/src/modules/proxyapp/proxyapp.controller.ts
@@ -1,10 +1,10 @@
 import { All, Controller, Param, Req, Res } from '@nestjs/common';
 import { ApiTags, ApiOperation, ApiParam } from '@nestjs/swagger';
-import { captureResponseBody } from '@beda.software/client-testing-proxy-helpers';
 import { RequestService } from '../requests/request.service';
 import { SessionService } from '../sessions/session.service';
 import { Request, Response } from 'express';
 import { createRequestObject } from '../../utils/data';
+import { captureResponseBody } from '../../utils/proxyHelpers';
 
 @ApiTags('proxy')
 @Controller('proxy')
