+++ packages/app/src/utils/data.spec.ts
@@ -1,4 +1,20 @@
-import { createTestListObject } from './data';
+import { createTestListObject, getFHIRAction } from './data';
+
+describe('getFHIRAction', () => {
+    test.each([
+        ['READ', 'GET', '/Patient/123'],
+        ['VREAD', 'GET', '/Patient/123/_history/1'],
+        ['CREATE', 'POST', '/Patient'],
+        ['UPDATE', 'PUT', '/Patient/123'],
+        ['DELETE', 'DELETE', '/Patient/123'],
+        ['SEARCH', 'GET', '/Patient?name=Smith'],
+        ['SEARCH', 'POST', '/Patient/_search'],
+        ['HISTORY', 'GET', '/Patient/_history'],
+        ['PATCH', 'PATCH', '/Patient/123'],
+    ])('should return correct FHIR action for %s', (expectedAction, method, url) => {
+        expect(getFHIRAction(method, url)).toBe(expectedAction);
+    });
+});
 
 describe('Data utilities', () => {
     it('Should return correct test list from test result object', async () => {
